Remove automatic door control actions.
This removes the automatic control of moving the door to a specified target which could lead to a dangerous situation.
The library should not take actions that the application may not expect, instead this should be handled directly by the app.

The app should instead determine the action to take, such as sounding a warning before beginning any closing motion or stopping
a moving door before setting the new target position.